% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpseg.R
\name{backtrace_r}
\alias{backtrace_r}
\title{backtracing \code{dpseg} segment break-points}
\usage{
backtrace_r(imax, jumps = 0)
}
\arguments{
\item{imax}{integer vector of segment borders as returned by
\code{dpseg} recursion functions}

\item{jumps}{allwo discontinuous jumps: move 1 index position back,
only for \eqn{S_{i-1} + score(i,j)}}
}
\value{
an integer vector of segment ends
}
\description{
Backtracing segment borders from the \code{imax} vector of a
\code{\link{dpseg}} recursion. This function is implemented more
efficiently in Rcpp; the R code is kept for documentation,
benchmarking and development.
}
