% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oddata}
\alias{oddata}
\title{Escherichia coli growth curves.}
\format{A data frame with the measurement time in column 1 and
    bacterial growth data (or blanks) in
    \code{2:ncol(oddata)}. Column names correspond to the well on
    the microtiter plate.}
\source{
Tom Rohr, Anna Behle, Rainer Machne, HHU Duesseldorf, 2018
}
\usage{
oddata
}
\description{
Optical density (OD) data from a 96-well microtiter plate experiment,
growing Escherichia coli cells in M9 medium in a BMG Optima platereader.
}
\keyword{datasets}
