% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpseg.R
\name{plot.dpseg}
\alias{plot.dpseg}
\title{Plot method for a \code{\link{dpseg}} segmentation model.}
\usage{
\method{plot}{dpseg}(x, delog = FALSE, col, main, ...)
}
\arguments{
\item{x}{result object returned by function \code{\link{dpseg}}}

\item{delog}{plot exp(y)}

\item{col}{optional color vector for segments}

\item{main}{plot title, dpseg parameters will be plotted if missing}

\item{...}{arguments passed to default \code{\link{plot}} function}
}
\value{
Silently returns the \code{x$segments} table , with color values
added if they were missing in the input.
}
\description{
Plot method for a \code{\link{dpseg}} segmentation model.
}
