% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpseg.R
\name{print.dpseg}
\alias{print.dpseg}
\title{Print method for linear segmentation result from \code{\link{dpseg}}.}
\usage{
\method{print}{dpseg}(x, ...)
}
\arguments{
\item{x}{result object returned by function \code{\link{dpseg}}}

\item{...}{further arguments to \code{print.data.frame}}
}
\description{
Prints the main result table \code{x$segments}, segment coordinates
and indices, and parameters from the recursion. See \code{\link{dpseg}}
for details.
}
